# openapi/bookops-api.openapi.yaml
openapi: 3.1.0
info:
  title: BookOps API
  version: 1.0.0
  description: >
    Thin HTTP API contract that maps directly to BookOps CLI operations and report artifacts.
servers:
  - url: /api
...
paths:
...
  /agent/run:
    post:
      operationId: runAgent
      tags: [Agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRunRequest'
      responses:
        '200':
          description: Agent execution result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRunEnvelope'
...
  /index/status:
    get:
      operationId: getIndexStatus
      tags: [Core]
      parameters:
        - name: include_symbolic
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Include full symbolic index entries in the response.
      responses:
        '200':
          description: Current index status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndexStatusEnvelope'
...
  /canon/diff:
    post:
      operationId: diffCanon
      tags: [Core]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CanonDiffRequest'
      responses:
        '200':
          description: Canon diff response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CanonDiffEnvelope'
  /analyze/chapter:
    post:
      operationId: analyzeChapter
      tags: [Analyze]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeChapterRequest'
      responses:
        '200':
          description: Chapter analysis summary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeChapterEnvelope'
  /analyze/project:
    post:
      operationId: analyzeProject
      tags: [Analyze]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeProjectRequest'
      responses:
        '200':
          description: Project analysis summary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeProjectEnvelope'
  /gate/chapter:
    post:
      operationId: gateChapter
      tags: [Gates]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GateChapterRequest'
      responses:
        '200':
          description: Gate result for chapter scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GateEnvelope'
  /gate/project:
    post:
      operationId: gateProject
      tags: [Gates]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GateProjectRequest'
      responses:
        '200':
          description: Gate result for project scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GateEnvelope'
  /pipeline/chapter:
    post:
      operationId: runChapterPipeline
      tags: [Pipeline]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PipelineChapterRequest'
      responses:
        '200':
          description: Chapter pipeline summary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineEnvelope'
  /pipeline/project:
    post:
      operationId: runProjectPipeline
      tags: [Pipeline]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PipelineProjectRequest'
      responses:
        '200':
          description: Project pipeline summary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineEnvelope'
  /issues:
    get:
      operationId: listIssues
      tags: [Issues]
      parameters:
        - $ref: '#/components/parameters/statusQuery'
        - $ref: '#/components/parameters/severityQuery'
        - $ref: '#/components/parameters/scopeQuery'
      responses:
        '200':
          description: Filtered issue list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueListEnvelope'
  /issues/{issueId}:
    patch:
      operationId: updateIssue
      tags: [Issues]
      parameters:
        - $ref: '#/components/parameters/issueIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueUpdateRequest'
      responses:
        '200':
          description: Updated issue status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueMutateEnvelope'
  /issues/{issueId}/waive:
    post:
      operationId: waiveIssue
      tags: [Issues]
      parameters:
        - $ref: '#/components/parameters/issueIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueWaiveRequest'
      responses:
        '200':
          description: Waived issue.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueMutateEnvelope'
  /lore/delta:
    post:
      operationId: generateLoreDelta
      tags: [Lore]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoreDeltaRequest'
      responses:
        '200':
          description: Lore delta summary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoreDeltaEnvelope'
  /lore/approve:
    post:
      operationId: approveLoreProposal
      tags: [Lore]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoreApproveRequest'
      responses:
        '200':
          description: Approved lore proposal.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoreApproveEnvelope'
  /lore/sync:
    post:
      operationId: applyLoreProposal
      tags: [Lore]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoreSyncRequest'
      responses:
        '200':
          description: Applied lore proposal.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoreSyncEnvelope'
  /reports/build:
    post:
      operationId: buildReports
      tags: [Reports]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportBuildRequest'
      responses:
        '200':
          description: Report build status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportBuildEnvelope'
  /reports/open:
    get:
      operationId: openReports
      tags: [Reports]
      parameters:
        - $ref: '#/components/parameters/reportScopeQuery'
        - $ref: '#/components/parameters/reportIdQuery'
      responses:
        '200':
          description: Report directory path.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportOpenEnvelope'
...
  /runs/{runId}:
    get:
      operationId: getRun
      tags: [Reports]
      parameters:
        - $ref: '#/components/parameters/runIdPath'
      responses:
        '200':
          description: One recorded pipeline run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunDetailEnvelope'
  /chapters/{chapterId}/content:
    get:
      operationId: getChapterContent
      tags: [Core]
      parameters:
        - $ref: '#/components/parameters/chapterIdPath'
      responses:
        '200':
          description: Chapter manuscript content.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterContentEnvelope'
...
  /settings:
    get:
      operationId: getSettings
      tags: [Core]
      responses:
        '200':
          description: Runtime settings payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsEnvelope'
    patch:
      operationId: patchSettings
      tags: [Core]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsPatchRequest'
      responses:
        '200':
          description: Updated runtime settings payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettingsEnvelope'
  /artifacts/chapter/{chapterId}/analysis:
    get:
      operationId: getChapterAnalysisArtifact
      tags: [Reports]
      parameters:
        - $ref: '#/components/parameters/chapterIdPath'
      responses:
        '200':
          description: Chapter analysis artifact.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactEnvelope'
  /artifacts/chapter/{chapterId}/gate:
    get:
      operationId: getChapterGateArtifact
      tags: [Reports]
      parameters:
        - $ref: '#/components/parameters/chapterIdPath'
      responses:
        '200':
          description: Chapter gate artifact.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactEnvelope'
  /artifacts/chapter/{chapterId}/decision-log:
    get:
      operationId: getChapterDecisionLogArtifact
      tags: [Reports]
      parameters:
        - $ref: '#/components/parameters/chapterIdPath'
      responses:
        '200':
          description: Chapter decision-log artifact.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactEnvelope'
  /artifacts/chapter/{chapterId}/continuity:
    get:
      operationId: getChapterContinuityArtifact
      tags: [Reports]
      parameters:
        - $ref: '#/components/parameters/chapterIdPath'
      responses:
        '200':
          description: Chapter continuity artifact.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactEnvelope'
  /artifacts/chapter/{chapterId}/style-audit:
    get:
      operationId: getChapterStyleAuditArtifact
      tags: [Reports]
      parameters:
        - $ref: '#/components/parameters/chapterIdPath'
      responses:
        '200':
          description: Chapter style-audit artifact.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactEnvelope'
  /artifacts/chapter/{chapterId}/lore-delta:
    get:
      operationId: getChapterLoreDeltaArtifact
      tags: [Reports]
      parameters:
        - $ref: '#/components/parameters/chapterIdPath'
      responses:
        '200':
          description: Chapter lore-delta artifact.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactEnvelope'
  /artifacts/chapter/{chapterId}/agent-results:
    get:
      operationId: getChapterAgentResultsArtifact
      tags: [Reports]
      parameters:
        - $ref: '#/components/parameters/chapterIdPath'
      responses:
        '200':
          description: Chapter agent-results artifact.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResultsEnvelope'
...
components:
  parameters:
    runIdPath:
      name: runId
      in: path
      required: true
      schema:
        type: string
    issueIdPath:
      name: issueId
      in: path
      required: true
      schema:
        type: string
    chapterIdPath:
      name: chapterId
      in: path
      required: true
      schema:
        type: integer
    statusQuery:
      name: status
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/IssueStatus'
    severityQuery:
      name: severity
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/IssueSeverity'
    scopeQuery:
      name: scope
      in: query
      required: false
      schema:
        type: string
    reportScopeQuery:
      name: scope
      in: query
      required: true
      schema:
        enum: [chapter, project]
    reportIdQuery:
      name: id
      in: query
      required: false
      schema:
        type: integer
  schemas:
    EnvelopeBase:
      type: object
      required: [ok, exit_code, data, stderr]
      properties:
        ok:
          type: boolean
        exit_code:
          type: integer
        data: {}
        stderr:
          type: string
...
    AgentRunRequest:
      type: object
      required: [agent_name, scope]
      properties:
        agent_name: { type: string }
        scope:
          type: string
          enum: [chapter, project]
        scope_id:
          type: integer
          nullable: true
    AgentResultData:
      type: object
      required: [name, summary, findings, proposals, confidence, needs_human_decision]
      properties:
        name: { type: string }
        summary: { type: string }
        findings:
          type: array
          items: {}
        proposals:
          type: array
          items: {}
        confidence: { type: number }
        needs_human_decision: { type: boolean }
    AgentRunEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AgentResultData'
    AgentResultsEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/AgentResultData'
...
    IndexStatusData:
      type: object
      required: [symbolic_exists, semantic_exists, symbolic_path, semantic_path]
      properties:
        symbolic_exists: { type: boolean }
        semantic_exists: { type: boolean }
        symbolic_path: { type: string }
        semantic_path: { type: string }
        index_version: { type: [integer, 'null'] }
        generated_at: { type: [string, 'null'] }
        file_count: { type: [integer, 'null'] }
        corpus_hash: { type: [string, 'null'] }
        symbolic:
          type: array
          items: { $ref: '#/components/schemas/IndexEntry' }
          description: Chapter/file entries when symbolic index exists
        semantic_status: { type: [string, 'null'] }
        semantic_source_file_count: { type: [integer, 'null'] }
    IndexStatusEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/IndexStatusData'
...
    CanonDiffRequest:
      type: object
      required: [from_snapshot, to_snapshot]
      properties:
        from_snapshot: { type: string }
        to_snapshot: { type: string }
    CanonDiffEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              type: object
              additionalProperties: true
    AnalyzeCheck:
      type: string
      enum: [tense, invariants, repetition, motifs, voice]
    AnalyzeChapterRequest:
      type: object
      required: [chapter_id]
      properties:
        chapter_id:
          type: integer
        checks:
          type: array
          items:
            $ref: '#/components/schemas/AnalyzeCheck'
        since:
          type: [string, 'null']
    AnalyzeChapterData:
      type: object
      required: [scope, findings, checks]
      properties:
        scope: { type: string }
        findings: { type: integer }
        checks:
          type: array
          items:
            $ref: '#/components/schemas/AnalyzeCheck'
        skipped: { type: [boolean, 'null'] }
        reason: { type: [string, 'null'] }
    AnalyzeChapterEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AnalyzeChapterData'
    AnalyzeProjectRequest:
      type: object
      properties:
        since:
          type: [string, 'null']
    AnalyzeProjectData:
      type: object
      required: [scope, findings, chapter_filter_count]
      properties:
        scope: { type: string }
        findings: { type: integer }
        chapter_filter_count: { type: integer }
    AnalyzeProjectEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AnalyzeProjectData'
...
    GateData:
      type: object
      required: [status, blocking_issue_ids, warning_issue_ids, message]
      properties:
        status:
          $ref: '#/components/schemas/GateStatus'
        blocking_issue_ids:
          type: array
          items: { type: string }
        warning_issue_ids:
          type: array
          items: { type: string }
        message:
          type: string
    GateChapterRequest:
      type: object
      required: [chapter_id]
      properties:
        chapter_id: { type: integer }
        strict: { type: boolean, default: false }
    GateProjectRequest:
      type: object
      properties:
        strict: { type: boolean, default: false }
    GateEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/GateData'
    PipelineChapterRequest:
      type: object
      required: [chapter_id]
      properties:
        chapter_id: { type: integer }
        strict: { type: boolean, default: false }
    PipelineProjectRequest:
      type: object
      properties:
        strict: { type: boolean, default: false }
    PipelineData:
      type: object
      required: [gate, scope]
      properties:
        gate:
          $ref: '#/components/schemas/GateStatus'
        scope:
          type: string
        run_id:
          type: string
          nullable: true
    PipelineEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/PipelineData'
    IssueSeverity:
      type: string
      enum: [critical, high, medium, low, info]
    IssueStatus:
      type: string
      enum: [open, in_progress, resolved, waived]
...
    IssueListData:
      type: object
      required: [count, issues]
      properties:
        count: { type: integer }
        issues:
          type: array
          items:
            $ref: '#/components/schemas/Issue'
    IssueListEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/IssueListData'
    IssueUpdateRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/IssueStatus'
        note:
          type: string
    IssueWaiveRequest:
      type: object
      required: [reason, reviewer]
      properties:
        reason: { type: string }
        reviewer: { type: string }
    IssueMutateData:
      type: object
      required: [issue_count]
      properties:
        updated: { type: [string, 'null'] }
        waived: { type: [string, 'null'] }
        issue_count: { type: integer }
    IssueMutateEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/IssueMutateData'
    LoreDeltaRequest:
      type: object
      required: [scope]
      properties:
        scope:
          type: string
          enum: [chapter, project]
        id:
          type: [integer, 'null']
        since:
          type: [string, 'null']
...
    LoreDeltaEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/LoreDeltaData'
    LoreApproveRequest:
      type: object
      required: [proposal, reviewer]
      properties:
        proposal: { type: string }
        reviewer: { type: string }
        note: { type: string }
...
    LoreApproveEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/LoreApproveData'
    LoreSyncRequest:
      type: object
      required: [proposal, apply]
      properties:
        proposal: { type: string }
        apply: { type: boolean }
...
    LoreSyncEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/LoreSyncData'
    ReportBuildRequest:
      type: object
      required: [scope]
      properties:
        scope:
          type: string
          enum: [chapter, project]
        id:
          type: [integer, 'null']
...
    ReportBuildEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ReportBuildData'
...
    ReportOpenEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ReportOpenData'
...
    RunDetailEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RunEntry'
...
    ChapterContentEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ChapterContentData'
...
    SettingsPatchRequest:
      type: object
      additionalProperties: true
    SettingsData:
      type: object
      additionalProperties: true
    SettingsEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/SettingsData'
...
    ArtifactEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ArtifactData'